chore: moved hooks folder
